backend/requirements.txt
python==3.11.5
fastapi==0.104.1
uvicorn==0.24.0
scapy==2.5.0
pyshark==0.6
python-dotenv==1.0.0 
sqlalchemy==2.0.23
websockets==12.0
scikit-learn==1.6.1
tensorflow==2.20.0
numpy==1.26.2
pandas==2.1.4
joblib==1.3.2
loguru==0.7.2
pytest==8.3.4
